+++ src/main/java/app/learningtrip/apiserver/category/domain/Grade.java
@@ -1,4 +1,4 @@
-package app.learningtrip.apiserver.course.domain;
+package app.learningtrip.apiserver.category.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -6,15 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+@Setter @Getter
 @NoArgsConstructor
-@Getter @Setter
 @Entity
-public class Province {
+public class Grade {
 
   @Id
   private Long id;
 
-  private String province;
+  private String school;
 
-  private String city;
+  private String grade;
 }
