+++ src/main/java/app/learningtrip/apiserver/category/domain/Province.java
@@ -1,4 +1,4 @@
-package app.learningtrip.apiserver.course.domain;
+package app.learningtrip.apiserver.category.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -6,15 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Setter @Getter
 @NoArgsConstructor
+@Getter @Setter
 @Entity
-public class Grade {
+public class Province {
 
   @Id
   private Long id;
 
-  private String school;
+  private String province;
 
-  private String grade;
+  private String city;
 }
