+++ src/main/java/app/learningtrip/apiserver/category/repository/GradeRepository.java
@@ -1,7 +1,6 @@
-package app.learningtrip.apiserver.course.repository;
+package app.learningtrip.apiserver.category.repository;
 
-import app.learningtrip.apiserver.course.domain.Grade;
-import java.util.List;
+import app.learningtrip.apiserver.category.domain.Grade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GradeRepository extends JpaRepository<Grade, Long> {
