+++ src/main/java/app/learningtrip/apiserver/category/repository/ProvinceRepository.java
@@ -1,6 +1,6 @@
-package app.learningtrip.apiserver.course.repository;
+package app.learningtrip.apiserver.category.repository;
 
-import app.learningtrip.apiserver.course.domain.Province;
+import app.learningtrip.apiserver.category.domain.Province;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProvinceRepository extends JpaRepository<Province, Long> {
