+++ src/main/java/app/learningtrip/apiserver/course/service/CourseService.java
@@ -5,11 +5,11 @@ import app.learningtrip.apiserver.course.domain.CourseRecommend;
 import app.learningtrip.apiserver.course.domain.CourseUser;
 import app.learningtrip.apiserver.course.domain.CoursePlaceUser;
 import app.learningtrip.apiserver.course.domain.GoogleMapApi;
-import app.learningtrip.apiserver.course.domain.Grade;
-import app.learningtrip.apiserver.course.domain.Province;
+import app.learningtrip.apiserver.category.domain.Grade;
+import app.learningtrip.apiserver.category.domain.Province;
 import app.learningtrip.apiserver.course.dto.request.CoursePlaceRequest;
 import app.learningtrip.apiserver.course.dto.request.CourseRequest;
-import app.learningtrip.apiserver.course.dto.response.CourseOptionsResponse;
+import app.learningtrip.apiserver.category.dto.CourseOptionsResponse;
 import app.learningtrip.apiserver.course.dto.response.CoursePlaceResponse;
 import app.learningtrip.apiserver.course.dto.response.CourseResponse;
 import app.learningtrip.apiserver.course.dto.response.CourseThumbnail;
@@ -17,8 +17,8 @@ import app.learningtrip.apiserver.course.repository.CoursePlaceRecommendReposito
 import app.learningtrip.apiserver.course.repository.CoursePlaceUserRepository;
 import app.learningtrip.apiserver.course.repository.CourseRecommendRepository;
 import app.learningtrip.apiserver.course.repository.CourseUserRepository;
-import app.learningtrip.apiserver.course.repository.GradeRepository;
-import app.learningtrip.apiserver.course.repository.ProvinceRepository;
+import app.learningtrip.apiserver.category.repository.GradeRepository;
+import app.learningtrip.apiserver.category.repository.ProvinceRepository;
 import app.learningtrip.apiserver.keyword.repository.KeywordRepository;
 import app.learningtrip.apiserver.place.domain.Place;
 import app.learningtrip.apiserver.place.repository.PlaceRepository;
@@ -52,14 +52,6 @@ public class CourseService {
     private final PlaceService placeService;
     private final PlaceRepository placeRepository;
     private final UserRepository userRepository;
-
-    private final GradeRepository gradeRepository;
-
-    private final ProvinceRepository provinceRepository;
-
-    private final KeywordRepository keywordRepository;
-
-    private final SearchServiceImpl searchService;
 
     private final GoogleMapApi googleMapApi;
 
@@ -352,30 +344,5 @@ public class CourseService {
                 }
             }
         }
-    }
-
-    /**
-     * 코스 옵션 (학년, 지역) 조회
-     * @return 코스 옵션 (학년, 지역)
-     */
-    public CourseOptionsResponse getCourseOptions(){
-        CourseOptionsResponse response = new CourseOptionsResponse();
-
-        List<Grade> gradeList = gradeRepository.findAll(Sort.by(Direction.ASC, "id"));
-        List<Province> provinceList = provinceRepository.findAll(Sort.by(Direction.ASC, "id"));
-
-        response.gradeToOptions(gradeList);
-        response.provinceToOptions(provinceList);
-
-        return response;
-    }
-
-    /**
-     * 코스 옵션 키워드 검색
-     * @param keyword 키워드
-     * @return 키워드 리스트
-     */
-    public List<String> getKeywordOfCourseOptions(String keyword){
-        return searchService.sortSearchList(keywordRepository.findKeywordsByNameLike(keyword), keyword);
     }
 }
